ワンパターンデメリットシステム(プレイヤー)は一応完成。 エネミーはまだ出来てない。 ※ただしObjFightのバグにより、 しっかりとしたバグチェック出来てない為、 プレイヤーの方にもバグ有る可能性有り。
▽以下、微修正。
podhpやm_player_damage等、
intよりfloatの方が都合が良いと思った変数をfloat化した。
初期化していないstatic変数あったので初期化。
コメント微修正。

+++ 星育喰/星育喰/GameHead.h
@@ -141,7 +141,7 @@ extern int g_Remain_num;			//残り住民数
 extern float g_Recast_time;			//ミサイルのリキャストタイム用
 //extern float g_Missile_pow;			//ミサイルの火力用
 //extern int g_E_Planet_damage;					//
-extern int g_P_Planet_damage;					//敵ポッドが惑星に与えるダメージ保存用
+extern float g_P_Planet_damage;					//敵ポッドが惑星に与えるダメージ保存用
 //extern int g_Enemy_Pod_LV;			//テスト用敵ポッドレベル
 //--------------------------------------------------資源----
 extern int g_Iron_num;				//鉄の所持数
